New Commands: Warning
You can now give a member a warning. When the amount of warnings a member has equals or gets bigger than 3, the member can be kicked. Usage: $givewarning @member (force). If you don't use force, it will just increase the amount of warns a member has. But if you use force, the bot will automatically kick the member.

Also added the operators (set, minus).

And a lot, really a lot of improvements. All commands have been updated. This can be bug fixes, or other additions.